Let Ingress be shared between blocking and async
