+++ atat/src/asynch/mod.rs
@@ -1,5 +1,3 @@
 mod client;
-mod ingress;
 
 pub use client::{AtatClient, Client};
-pub use ingress::{AtatIngress, Ingress};
