+++ atat/src/buffers.rs
@@ -1,6 +1,6 @@
 use bbqueue::BBBuffer;
 
-use crate::{Config, Digester};
+use crate::{Config, Digester, Ingress};
 
 pub struct Buffers<
     const INGRESS_BUF_SIZE: usize,
@@ -36,41 +36,41 @@ impl<const INGRESS_BUF_SIZE: usize, const RES_CAPACITY: usize, const URC_CAPACIT
         digester: D,
         config: Config,
     ) -> (
-        crate::asynch::Ingress<'a, D, INGRESS_BUF_SIZE, RES_CAPACITY, URC_CAPACITY>,
+        Ingress<'a, D, INGRESS_BUF_SIZE, RES_CAPACITY, URC_CAPACITY>,
         crate::asynch::Client<'a, W, Clk, Delay, RES_CAPACITY, URC_CAPACITY>,
+    ) {
+
+        let (res_writer, res_reader) = self.res_queue.try_split_framed().unwrap();
+        let (urc_writer, urc_reader) = self.urc_queue.try_split_framed().unwrap();
+
+        (
+            Ingress::new(digester, res_writer, urc_writer),
+            crate::asynch::Client::new(writer, clock, delay, res_reader, urc_reader, config),
+        )
+    }
+
+    pub fn split_blocking<
+        'a,
+        W: embedded_hal_nb::serial::Write<u8>,
+        Tim: fugit_timer::Timer<TIMER_HZ>,
+        D: Digester,
+        const TIMER_HZ: u32
+    >(
+        &'a self,
+        writer: W,
+        timer: Tim,
+        digester: D,
+        config: Config,
+    ) -> (
+        Ingress<'a, D, INGRESS_BUF_SIZE, RES_CAPACITY, URC_CAPACITY>,
+        crate::blocking::Client<'a, W, Tim, TIMER_HZ, RES_CAPACITY, URC_CAPACITY>,
     ) {
         let (res_writer, res_reader) = self.res_queue.try_split_framed().unwrap();
         let (urc_writer, urc_reader) = self.urc_queue.try_split_framed().unwrap();
 
         (
-            crate::asynch::Ingress::new(digester, res_writer, urc_writer),
-            crate::asynch::Client::new(writer, clock, delay, res_reader, urc_reader, config),
+            Ingress::new(digester, res_writer, urc_writer),
+            crate::blocking::Client::new(writer, res_reader, urc_reader, timer, config),
         )
     }
-
-    // pub fn split_blocking<
-    //     'a,
-    //     W: embedded_io::blocking::Write,
-    //     Clk: embedded_time::Clock,
-    //     Delay: embedded_hal_async::delay::DelayUs,
-    //     D: Digester,
-    // >(
-    //     &'a self,
-    //     writer: W,
-    //     clock: &'a Clk,
-    //     delay: Delay,
-    //     digester: D,
-    //     config: Config,
-    // ) -> (
-    //     crate::blocking::Ingress<'a, D, INGRESS_BUF_SIZE, RES_CAPACITY, URC_CAPACITY>,
-    //     crate::blocking::Client<'a, W, Clk, Delay, RES_CAPACITY, URC_CAPACITY>,
-    // ) {
-    //     let (res_writer, res_reader) = self.res_queue.try_split_framed().unwrap();
-    //     let (urc_writer, urc_reader) = self.urc_queue.try_split_framed().unwrap();
-
-    //     (
-    //         crate::blocking::Ingress::new(digester, res_writer, urc_writer),
-    //         crate::blocking::Client::new(writer, clock, delay, res_reader, urc_reader, config),
-    //     )
-    // }
 }
